Update Othello.hs
added a random number generator. it will generate random number but you will have to call r1 <-randomize everytime to get a new number but the issues is that I can't figure out how to plus it into the function jasonStrategy then print the board. 
basically what I have in mind is 
initial move 
then 
randomize move if it is valid play (flip or not) if not quit <---have not figure out that part yet